Adding a single `aiohttp` test.
