+++ tests/main/test_asyncio.py
@@ -4,7 +4,11 @@ import json
 import socket
 import tempfile
 
-from mocket import Mocketizer
+import aiohttp
+import pytest
+
+from mocket import Mocketizer, async_mocketize
+from mocket.mockhttp import Entry
 
 
 def test_asyncio_record_replay(event_loop):
@@ -37,3 +41,21 @@ def test_asyncio_record_replay(event_loop):
             responses = json.load(f)
 
         assert len(responses["google.com"]["80"].keys()) == 1
+
+
+@pytest.mark.asyncio
+@async_mocketize
+async def test_aiohttp():
+    url = "https://bar.foo/"
+    data = {"message": "Hello"}
+
+    Entry.single_register(
+        Entry.GET,
+        url,
+        body=json.dumps(data),
+        headers={"content-type": "application/json"},
+    )
+
+    async with aiohttp.ClientSession() as session, session.get(url) as response:
+        response = await response.json()
+        assert response == data
